Fixed CreatePost issues
