+++ frontend/src/pages/Feed/FeedPage.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 
 import { getPosts } from "../../services/posts";
 import Post from "../../components/Post/Post";
-import CreatePost from "../../components/CreatePost/CreatePost";
+import CreatePost from "../../components/Post/CreatePost";
 
 export const FeedPage = () => {
   const [posts, setPosts] = useState([]);
@@ -24,7 +24,7 @@ export const FeedPage = () => {
     } else {
       navigate("/login");
     }
-  });
+  }, [posts]);
 
   if (!token) {
     return;
@@ -32,7 +32,8 @@ export const FeedPage = () => {
 
   return (
     <>
-      <CreatePost />
+      <CreatePost token={token}/>
+
       <h2>Posts</h2>
       <div className="feed" role="feed">
         {posts.map((post) => (
